Initial commit : Start of the project
Annotations : created the basic annotations. Without any fields. Just to make a standard processor work with them

Processor : Created a basic processor that generates an implementation for each class that extends SearchableRepository. The generated class produces a method with the name of all the fields that are annotated by @Searchable just to make sure our processor works.