Switch from generating the implementation if an interface extends our "Repository" to generating directly from @SeachableEntity.
Why ?
Interface was made so our IDE can auto complete, turns out the interface is empty so there are no auto completion using it.

Right now we have to actually compile the project to generate our classes. Could be something to improve in the future ? lombok-like ? We'll see.

Added a readme just for the pom.xml dependencies. Will complete it somedays maybe